// libs/config/src/enums/Step.ts
export enum Step {
  findPackages = "findPackages",
  verifyConditions = "verifyConditions",
  analyzeCommits = "analyzeCommits",
  verifyRelease = "verifyRelease",
  generateNotes = "generateNotes",
  addChannels = "addChannels",
  prepare = "prepare",
  publish = "publish",
  success = "success",
  fail = "fail",
}